remove pending message (auto register pending key)
