+++ client-g2/src/client/session-reducer.ts
@@ -187,7 +187,6 @@ export function reduceMessage(
 
     case "auth.error":
     case "auth.challenge":
-    case "pairing.pending":
     case "pong":
       return null;
   }
